Fix up Docks::Containers
