+++ lib/docks/containers/class_container.rb
@@ -1,4 +1,4 @@
-require File.expand_path("../base_container.rb", __FILE__)
+require_relative "base_container.rb"
 
 module Docks
   module Containers
